reworking the structure
